fix: 기존 Detail 페이지 Search로 변경
