+++ packages/react-native/src/routes/HomeStackNavigator.tsx
@@ -1,8 +1,10 @@
 import { createStackNavigator } from '@react-navigation/stack';
 import Home from '@/pages/Home';
+import { StackParamList } from '@/types/navigation';
+import Search from '@/pages/Search';
 import Detail from '@/pages/Detail';
 
-const Stack = createStackNavigator();
+const Stack = createStackNavigator<StackParamList>();
 
 export default function HomeStackNavigator() {
   return (
@@ -11,6 +13,7 @@ export default function HomeStackNavigator() {
       initialRouteName="Home/Main"
     >
       <Stack.Screen name="Home/Main" component={Home} />
+      <Stack.Screen name="Home/Search" component={Search} />
       <Stack.Screen name="Home/Detail" component={Detail} />
     </Stack.Navigator>
   );
